move: ResourceReader 클래스 패키지 response 하위로 이동
