+++ tomcat/src/main/java/org/apache/coyote/http11/response/ResourceReader.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.response;
 
 import java.io.File;
 import java.io.IOException;
